filter now applies case-insensitive

+++ src/events/message.ts
@@ -5,19 +5,30 @@ module.exports = {
 	name: Events.MessageCreate,
 	once: false,
 	execute(message: Message) {
+		if (message.author.tag === message.client.user.tag) {
+			return
+		}
 		let low = message.toString().toLowerCase()
 		let words = low.replace(/[.,\/?#!$%\^&\*;:{}=\-_`~()]/g,"").split(" ")
+		let skip = false
 		filter.forEach(async (word: string) => {
-			if (words.includes(word)) {
+			if (!skip && words.includes(word)) {
+				skip = true
 				try {
 					const fetchedMessage = await message.channel.messages.fetch(message.id);
 					await fetchedMessage.delete();
-					console.log(`Deleted message with ID ${message.id}`);
-					message.channel.send("Message deleted")
+
+					let str = message.toString()
+					filter.forEach(w => {
+						var regEx = new RegExp(w, "ig");
+						str = str.toString().replaceAll(regEx, "[REDACTED]");
+					})
+					
+					await message.channel.send(str)
+
 				} catch (error) {
 					console.error(`Failed to delete message with ID ${message.id}: ${error}`);
 				}
-				return
 			}
 		})
 	},
